Move build_assoc and build_cons to pyk.kast
